(cherry picked from commit 437cd7cf310cfdc3cfdbaaead4a62ce7e75cf118)
(cherry picked from commit fdaaf0521d8c55edde48fd76976fcb051e354375)

+++ src/main/scala/org/bitlap/zim/repository/ReceiveRepository.scala
@@ -11,7 +11,7 @@ import zio.stream
  */
 trait ReceiveRepository[T] extends BaseRepository[T] {
 
-  def saveMessage(receive: T): stream.Stream[Throwable, Long]
+  def saveMessage(receive: T): stream.Stream[Throwable, Int]
 
   def findOffLineMessage(uid: Int, status: Int): stream.Stream[Throwable, T]
